requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
groq
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-text-splitters
pypdf
python-multipart
rank_bm25
passlib
bcrypt
PyJWT
psycopg2-binary
supabase
langchain-postgres
